templates/telegram-bot/requirements.txt
python-telegram-bot[rate-limiter]==21.6
python-dotenv==1.0.1
